+++ quic/s2n-quic-transport/src/endpoint/mod.rs
@@ -120,10 +120,9 @@ impl<Cfg: Config> s2n_quic_core::endpoint::Endpoint for Endpoint<Cfg> {
             }
         }
 
-        // process ACKs on Connects with interest
-        self.connections.iterate_new_ack_list(|connection| {
+        // process ACKs on Connections with interest
+        self.connections.iterate_ack_list(|connection| {
             connection.on_process_acks();
-            ConnectionContainerIterationResult::Continue
         });
 
         let len = entries.len();
